Fold the two printers into one with a strategy switch.
This is another pattern. Much more tightly coupled, but given the amount
of shared stuff in this output, this is another way you could go.
