+++ log_printer.rb
@@ -4,31 +4,50 @@
 require './log_processor'
 
 class LogPrinter
-  attr_accessor :page_stats
+  attr_reader :page_stats, :strategy
 
-  def initialize(page_stats)
+  def initialize(page_stats, strategy: :list)
     @page_stats = page_stats
+    @strategy = strategy
   end
 
   def to_s
     [
       'WEBPAGES WITH MOST PAGE VIEWS:',
-      fmt_list(:page_views, 'visits'),
+      fmt(:page_views, 'visits'),
       'WEBPAGES WITH MOST UNIQUE PAGE VIEWS:',
-      fmt_list(:unique_page_views, 'unique views')
-    ].join("\n")
+      fmt(:unique_page_views, 'unique views')
+    ].join("\n\n")
   end
 
   private
+
+  def fmt(key, label)
+    case strategy
+    when :table
+      fmt_table(key, label)
+    else
+      fmt_list(key, label)
+    end
+  end
 
   def fmt_list(key, count_label)
     hash = page_stats.transform_values{|stats| stats.send(key)}.sort_by(&pair_arr_sort).reverse
     hash.map { |path, value| "#{path} - #{value} #{count_label}" }.join("\n")
   end
 
-  def pair_arr_sort 
-    lambda {|(k, v)| v }
+  def fmt_table(key, value_label)
+    hash = page_stats.transform_values{|stats| stats.send(key)}.sort_by(&pair_arr_sort).reverse
+    key_max_size = hash.to_h.keys.map(&:size).max
+    value_max_size = hash.to_h.values.map(&:to_s).map(&:size).max
+    hash.map do |path, uniq_ips|
+      "| %#{key_max_size}s | %#{value_max_size}s %s |" %
+        [path, uniq_ips, value_label]
+    end.join("\n")
   end
 
+  def pair_arr_sort
+    lambda {|(k, v)| v }
+  end
 
 end
